Python/cloud_functions/Insert_Int_Func/src/requirements.txt
google-cloud-secret-manager==2.12.6
google-cloud-bigquery==3.5.0
python-dotenv==0.21.0
pandas==1.5.3
pyarrow==11.0.0